// src/types/Movie.ts
export interface Movie {
    imdbID: string;
    Title: string;
    Year: string;
    Poster: string;
    Genre: string;
    imdbRating: string;
    Released: string;
    Runtime: string;
    Director: string;
    Writer: string;
    Actors: string;
    BoxOffice: string;
    Production: string;
    Country: string;
    Plot: string;
}